+++ srcjs/bs4Dash-2.2.2.9000/miscellaneous.js
@@ -115,7 +115,8 @@ $(function () {
 
 
   // tooltip/popover toggle
-  if ($('body').attr('data-help') == 1) {
+  if ($('body').attr('data-help') == 2 || 
+    $('body').attr('data-help') == 1) {
     var $help_switch_checkbox = $('<input />', {
       type: 'checkbox',
       id: 'help_switch',
@@ -143,7 +144,11 @@ $(function () {
     }
 
     // trigger first click, if necessary
-    $help_switch_checkbox.click();
+    if ($('body').attr('data-help') == 2) {
+      $(document).on('shiny:connected', function() {
+        $help_switch_checkbox.click();
+      }); 
+    }
   }
 
   // dark mode input
@@ -232,7 +237,7 @@ $(function () {
   $('body').attr('data-dark') == 1) {
     var $dark_mode_checkbox = $('<input />', {
       type: 'checkbox',
-      id: 'customSwitch1',
+      id: 'theme_switch',
       class: 'custom-control-input'
     }).on('click', function () {
 
@@ -344,7 +349,7 @@ $(function () {
     });
 
     var $dark_mode_icon = $('body').hasClass('dark-mode') ? '<i class="dark-theme-icon fa fa-moon"></i>' : '<i class="dark-theme-icon fa fa-sun"></i>';
-    var $dark_mode_container = $('<div />', { class: 'custom-control custom-switch mx-2 mt-2' }).append($dark_mode_checkbox).append(`<label class="custom-control-label" for="customSwitch1">${$dark_mode_icon}</label>`);
+    var $dark_mode_container = $('<div />', { class: 'custom-control custom-switch mx-2 mt-2' }).append($dark_mode_checkbox).append(`<label class="custom-control-label" for="theme_switch">${$dark_mode_icon}</label>`);
     
     // insert before $('#controlbar-toggle') whenever possible ...
     if ($('.nav-item #controlbar-toggle')) {
@@ -357,7 +362,7 @@ $(function () {
     // Trigger dark mode
     if ($('body').attr('data-dark') == 2) {
       $(document).on('shiny:connected', function() {
-        $('#customSwitch1').click();
+        $dark_mode_checkbox.click();
       }); 
     }
   }